Building 2020, calibrated with new traditional biomass number
By following the methodology used to calculate traditional biomass numbers, it could be seen that the number was wrong. The new number is calculated using the NFHS 5 Report. Traditional biomass numbers were reduced to 1/3rd of the value previously used to calibrate the building's energy demand. Earlier, its share was around 80-90% of total energy, which has now reduced to around 50%.